requirements.txt
click>=8.0.0
colorama>=0.4.4
pyyaml>=6.0
openai>=0.27.0 